chore(refactor): various module/folder restructure - moved out "typeRegistry" to seperate module - moved annotations to seperate folder
